fixed cyclic imports, updated and fixed documentation
